llama_service/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic 